modify the subclasses
